test pasing

+++ arcus/exc.py
@@ -85,7 +85,7 @@ class RecurrentPayments(UnprocessableEntity):
 
 
 class DuplicatedPayment(UnprocessableEntity):
-    def __init__(self, code: str, amount: int):
+    def __init__(self, code: str, amount: float):
         message = f'Duplicated payment for {amount}'
         super().__init__(
             code, message
@@ -93,7 +93,7 @@ class DuplicatedPayment(UnprocessableEntity):
 
 
 class IncompleteAmount(UnprocessableEntity):
-    def __init__(self, code: str, amount: int):
+    def __init__(self, code: str, amount: float):
         message = (
             f'Incomplete payment amount of {amount}, must pay full balance')
         super().__init__(
